Restructure files to match upstream structure for v1.3
This is in preparation for upgrading to v1.3 as part of the
work to upgrade mockito to 1.10.19.

src/main -> dexmaker/src/main
src/test -> dexmaker/src/test
src/dx -> dx/src/main
src/mockito -> mockito/src/main

The README is moved to README.google to be consistent with other
open source based projects in AOSP.

Also updates the patch files and instructions on how to use
them.

Bug: 30299479
Test: Build and run some mockito based tests.
Change-Id: I2d80acc1e117a26e55dc8c4522eb9bc79f3852cc