Revert upstream change 924bb5b1c1718ac4c49d94e435739a6fd26dc986
The reverted change broke some android tests because it did not
traverse the interface inheritance hierarchy when searching for
methods to proxy.

That change was introduced during the upgrade to 1.3 and so
while it is fixing a problem none of the Android code that uses
this version of dexmaker was affected by it so it should be safe
to remove.

Will work on a proper fix and test for upstream after this has
been reverted as this is blocking other teams at the moment.

Bug: 30299479
Change-Id: Id4a7dacdf74ac2b08ca28a7677b71b191b0c36bd
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh