Reapply - Hacks to allow mocking of package private classes
2 things are required to allow mocking for package private classes.
First is creating the mock in the same class as the base, that part
is simple and seems to have no effect. The second is to have a matching
ClassLoader as the source class.

To make the classloader match, instead of generating a new ClassLoader,
add the new dex to the existing one and share the base class's ClassLoader.
Since this is a persistent change and could affect the process, it will be
opt-in through setting a system property.

This also appears to have unintended speedup in the tests I was testing,
possible through cache fixes or brokenness.

(cherry picked from commit 31b974435499d3c9769ed6a746e946187f3be64f)

Test: Add System.setProperty("dexmaker.share_classloader", "true) to SysuiTestCase
    then runtest systemui

Change-Id: I35c805ee3797ff151ac19bd0c931ebfd41d3b670